Round loss percentage to 2 decimal places in the table.

+++ src/main/java/app/Country.java
@@ -14,7 +14,7 @@ public class Country {
    private String name;
 
    public String year;
-   public int lossPercent;
+   public double lossPercent;
    public String activity;
    public String cause;
    public String supplyStage;
